requirements.txt
biopython @ file:///Users/runner/miniforge3/conda-bld/biopython_1737241778882/work
imbalanced-learn @ file:///home/conda/feedstock_root/build_artifacts/imbalanced-learn_1734717084342/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1733736026804/work
numpy==2.2.3
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1736496824048/work/dist/scikit_learn-1.6.1-cp312-cp312-macosx_11_0_arm64.whl#sha256=b73b4426f1e03a197903eddede31a62b6859e1eb5d298b145d003d742720c9f7
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1736617284906/work/dist/scipy-1.15.1-cp312-cp312-macosx_11_0_arm64.whl#sha256=e9ceeea8211d7cfb6dcfd39a9ef55b9c577d449365006ea3219f89dc86839813
setuptools==75.8.0
sklearn-compat @ file:///home/conda/feedstock_root/build_artifacts/sklearn-compat_1734794177910/work
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1714400101435/work
wheel==0.45.1
